Revert "Revert "People picker refactor" (#1142)"
This reverts commit c103e18facc5e3e1308a4916850a1a738556ba33.
