+++ app/components/pages/SubmissionWizard/steps/Editors/EditorsPage.js
@@ -4,7 +4,7 @@ import { Box } from 'grid-styled'
 import { Action, ErrorText } from '@pubsweet/ui'
 import { FormH3 } from '../../../../ui/atoms/FormHeadings'
 
-import PeoplePickerControl from './PeoplePickerControl'
+import PeoplePickerControl from '../../../../ui/PeoplePicker/PeoplePickerControl'
 import Textarea from '../../../../ui/atoms/Textarea'
 import CalloutBox from '../../../../ui/atoms/CalloutBox'
 import ValidatedField from '../../../../ui/atoms/ValidatedField'
@@ -150,6 +150,7 @@ class EditorsPage extends React.Component {
               initialSelection={values.suggestedSeniorEditors}
               maxSelection={limits.suggestedSeniorEditors.max}
               minSelection={limits.suggestedSeniorEditors.min}
+              modalName="suggestedSeniorEditors"
               onRequestRemove={person =>
                 this.removeSelection('suggestedSeniorEditors', person)
               }
@@ -179,6 +180,7 @@ class EditorsPage extends React.Component {
               initialSelection={values.opposedSeniorEditors}
               maxSelection={limits.opposedSeniorEditors.max}
               minSelection={limits.opposedSeniorEditors.min}
+              modalName="opposedSeniorEditors"
               onRequestRemove={person =>
                 this.removeSelection('opposedSeniorEditors', person)
               }
@@ -211,6 +213,7 @@ class EditorsPage extends React.Component {
               initialSelection={values.suggestedReviewingEditors}
               maxSelection={limits.suggestedReviewingEditors.max}
               minSelection={limits.suggestedReviewingEditors.min}
+              modalName="suggestedReviewingEditors"
               onRequestRemove={person =>
                 this.removeSelection('suggestedReviewingEditors', person)
               }
@@ -242,6 +245,7 @@ class EditorsPage extends React.Component {
               initialSelection={values.opposedReviewingEditors}
               maxSelection={limits.opposedReviewingEditors.max}
               minSelection={limits.opposedReviewingEditors.min}
+              modalName="opposedReviewingEditors"
               onRequestRemove={person =>
                 this.removeSelection('opposedReviewingEditors', person)
               }
