+++ app/components/ui/PeoplePicker/ChooserPod.md
@@ -1,0 +1,15 @@
+```js
+;<ChooserPod
+  roleName="Senior Editor(s)"
+  isRequired={true}
+  togglePersonSelection={() => console.log('icon clicked')}
+/>
+```
+
+```js
+;<ChooserPod
+  roleName="Reviewing Editor(s)"
+  isRequired={false}
+  togglePersonSelection={() => console.log('icon clicked')}
+/>
+```
