+++ app/components/ui/PeoplePicker/PersonPod.md
@@ -1,5 +1,3 @@
-### ChosenPersonPod
-
 Clickable keywords:
 
 ```js
@@ -12,7 +10,7 @@ Clickable keywords:
   onKeywordClick={keyword => console.log(keyword, 'clicked')}
   isPicked={false}
   isStatusShown={false}
-  iconType="add"
+  selectButtonType="add"
   togglePersonSelection={() => console.log('icon clicked')}
 />
 ```
@@ -29,7 +27,7 @@ Showing the person's status:
   status="Currently unavailable"
   isPicked={false}
   isStatusShown={true}
-  iconType="add"
+  selectButtonType="add"
   togglePersonSelection={() => console.log('icon clicked')}
 />
 ```
@@ -47,7 +45,7 @@ Removal icon:
   status="Currently unavailable"
   isSelectedInForm={false}
   isStatusShown={true}
-  iconType="remove"
+  selectButtonType="remove"
   togglePersonSelection={() => console.log('icon clicked')}
 />
 ```
@@ -63,7 +61,7 @@ Selected icon:
   isKeywordClickable={false}
   isSelectedInForm={false}
   isStatusShown={true}
-  iconType="selected"
+  selectButtonType="selected"
   togglePersonSelection={() => console.log('icon clicked')}
 />
 ```
@@ -77,27 +75,9 @@ Disabled:
   focuses={['biological experiments']}
   expertises={['Cell Biology']}
   isKeywordClickable={false}
-  iconType="add"
+  selectButtonType="add"
   isSelectedButtonClickable={false}
   isSelectedInForm={false}
-  togglePersonSelection={() => console.log('icon clicked')}
-/>
-```
-
-### SelectButton
-
-```js
-;<PersonPod.SelectButton
-  roleName="Senior Editor(s)"
-  isRequired={true}
-  togglePersonSelection={() => console.log('icon clicked')}
-/>
-```
-
-```js
-;<PersonPod.SelectButton
-  roleName="Reviewing Editor(s)"
-  isRequired={false}
   togglePersonSelection={() => console.log('icon clicked')}
 />
 ```
