--- app/components/ui/atoms/PersonInfo.md
@@ -1,13 +1,0 @@
-```js
-<PersonInfo
-  id="1234"
-  name="Michael J Frank"
-  institution="Max Planck Institute, Germany"
-  focuses={[
-    'genetics of global gene expression',
-    'mouse models of human disease',
-    'haematology',
-  ]}
-  expertises={['Neuroscience', 'Cell Biology']}
-/>
-```
